backend/pyproject.toml
[project]
name = "sempkm"
version = "0.1.0"
description = "SemPKM - Semantics-native Personal Knowledge Management"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=3.0",
    "fastapi[standard]",
    "rdflib>=7.5.0",
    "httpx>=0.28",
    "pydantic-settings",
    "jinja2",
    "jinja2-fragments",
    "uvicorn",
    "pyshacl>=0.31.0",
    "cachetools>=7.0",
    "pyyaml>=6.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "alembic>=1.18",
    "aiosqlite>=0.22",
    "asyncpg>=0.31",
    "itsdangerous>=2.2",
    "cryptography>=43.0",
    "posthog>=3.7",
    "wsgidav>=4.3.3,<5.0",
    "a2wsgi>=1.10",
    "python-frontmatter>=1.1.0",
    "owlrl>=7.0",
    "mf2py>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
